refactor(yyzeiluo): update folder name
