+++ app/[lang]/page.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import ArticlePopular from "@/components/article-popular";
-import AboutMyTeam from "@/components/m-home/about-my-team";
-import Category from "@/components/m-home/category";
-import HeroBanner from "@/components/m-home/hero-banner";
-import News from "@/components/m-home/news";
+import ArticlePopular from "@/components/features/article-popular";
+import AboutMyTeam from "@/components/features/home/about-my-team";
+import Category from "@/components/features/home/category";
+import HeroBanner from "@/components/features/home/hero-banner";
+import News from "@/components/features/home/news";
 
 import { getDictionary } from "@/dictionaries/get-dictionary";
 import { fetchCategories } from "@/services/category";
