+++ components/features/article-popular/editor-choices/index.tsx
@@ -1,15 +1,17 @@
 import React from "react";
 import isEqual from "react-fast-compare";
 
-import AppLink from "@/components/common/app-link";
 import AppSectionTitle from "@/components/common/app-section-title";
+import CardSingleBlog from "@/components/features/card-blog-item";
 
-import { randomId } from "@/lib/randomId";
+import { randomId } from "@/lib/random-id";
+import { cn } from "@/lib/utils";
 import { LocaleKeys } from "@/types/locales";
 
 interface Props {
   dictionary: LocaleKeys;
 }
+
 const data = [
   {
     title:
@@ -19,6 +21,9 @@ const data = [
     image:
       "https://i.pinimg.com/564x/1c/34/0c/1c340cf030e39aa9d924e9b0418238ec.jpg",
     id: randomId(),
+    postingTime: "Dec 12",
+    readingTime: "5 min read",
+    author: "jake",
     href: "/",
   },
   {
@@ -29,6 +34,9 @@ const data = [
     image:
       "https://i.pinimg.com/564x/1c/34/0c/1c340cf030e39aa9d924e9b0418238ec.jpg",
     id: randomId(),
+    author: "jake",
+    postingTime: "Dec 12",
+    readingTime: "5 min read",
     href: "/",
   },
   {
@@ -36,45 +44,41 @@ const data = [
       " Nearly 200 Great Barrier Reef coral species also live in the deep sea",
     description:
       "There are more coral species lurking in the deep ocean that previously throught.",
-    image:
-      "https://i.pinimg.com/564x/1c/34/0c/1c340cf030e39aa9d924e9b0418238ec.jpg",
+    image: "https://images7.alphacoders.com/129/thumb-440-1299888.webp",
     id: randomId(),
-    href: "/",
+    postingTime: "Dec 12",
+    readingTime: "5 min read",
+    author: "jake",
+    href: "/3",
   },
 ];
 
-function Popular(props: Props) {
+function EditorChoices(props: Props) {
   const { dictionary } = props;
   return (
     <>
       <AppSectionTitle
         dictionary={dictionary}
-        title="Popular"
+        title="All Stories"
         linkShowMore="/allNews"
         className="mb-8"
       />
 
-      <div className="flex flex-col gap-3">
-        {data.map((item, index) => (
-          <article key={item.id} className="flex justify-center gap-4">
-            <span className="flex text-center text-big-size font-bold">
-              {index < 9 ? `0${index + 1}` : index + 1}
-            </span>
-
-            <div className="mt-2 flex flex-1 flex-col justify-center">
-              <h3 className="line-clamp-1 text-heading-5 font-bold leading-7">
-                <AppLink href={item.href} aria-label={item.title}>
-                  Did Supernovae
-                </AppLink>
-              </h3>
-
-              <span className="line-clamp-2">Jake Bittle in SCIENCE</span>
-            </div>
-          </article>
-        ))}
-      </div>
+      {data.map((blog, index) => (
+        <CardSingleBlog
+          key={blog.id}
+          author={blog.author}
+          description={blog.description}
+          href={blog.href}
+          image={blog.image}
+          postingTime={blog.postingTime}
+          readingTime={blog.readingTime}
+          title={blog.title}
+          className={cn(!!index && "mt-4")}
+        />
+      ))}
     </>
   );
 }
 
-export default React.memo(Popular, isEqual);
+export default React.memo(EditorChoices, isEqual);
