+++ components/features/article-popular/popular/index.tsx
@@ -1,16 +1,15 @@
 import React from "react";
+import isEqual from "react-fast-compare";
 
-import CardSingleBlog from "@/components/card-blog-item";
+import AppLink from "@/components/common/app-link";
 import AppSectionTitle from "@/components/common/app-section-title";
 
-import { randomId } from "@/lib/randomId";
-import { cn } from "@/lib/utils";
+import { randomId } from "@/lib/random-id";
 import { LocaleKeys } from "@/types/locales";
 
 interface Props {
   dictionary: LocaleKeys;
 }
-
 const data = [
   {
     title:
@@ -20,9 +19,6 @@ const data = [
     image:
       "https://i.pinimg.com/564x/1c/34/0c/1c340cf030e39aa9d924e9b0418238ec.jpg",
     id: randomId(),
-    postingTime: "Dec 12",
-    readingTime: "5 min read",
-    author: "jake",
     href: "/",
   },
   {
@@ -33,9 +29,6 @@ const data = [
     image:
       "https://i.pinimg.com/564x/1c/34/0c/1c340cf030e39aa9d924e9b0418238ec.jpg",
     id: randomId(),
-    author: "jake",
-    postingTime: "Dec 12",
-    readingTime: "5 min read",
     href: "/",
   },
   {
@@ -46,40 +39,42 @@ const data = [
     image:
       "https://i.pinimg.com/564x/1c/34/0c/1c340cf030e39aa9d924e9b0418238ec.jpg",
     id: randomId(),
-    postingTime: "Dec 12",
-    readingTime: "5 min read",
-    author: "jake",
-    href: "/3",
+    href: "/",
   },
 ];
 
-function News(props: Props) {
+function Popular(props: Props) {
   const { dictionary } = props;
-
   return (
-    <section className="container my-10">
+    <>
       <AppSectionTitle
         dictionary={dictionary}
-        title="All Stories"
+        title="Popular"
         linkShowMore="/allNews"
         className="mb-8"
       />
 
-      {data.map((blog, index) => (
-        <CardSingleBlog
-          key={blog.id}
-          author={blog.author}
-          description={blog.description}
-          href={blog.href}
-          image={blog.image}
-          postingTime={blog.postingTime}
-          readingTime={blog.readingTime}
-          title={blog.title}
-          className={cn(!!index && "mt-4")}
-        />
-      ))}
-    </section>
+      <div className="flex flex-col gap-3">
+        {data.map((item, index) => (
+          <article key={item.id} className="flex justify-center gap-4">
+            <span className="flex text-center text-big-size font-bold">
+              {index < 9 ? `0${index + 1}` : index + 1}
+            </span>
+
+            <div className="mt-2 flex flex-1 flex-col justify-center">
+              <h3 className="line-clamp-1 text-heading-5 font-bold leading-7">
+                <AppLink href={item.href} aria-label={item.title}>
+                  Did Supernovae
+                </AppLink>
+              </h3>
+
+              <span className="line-clamp-2">Jake Bittle in SCIENCE</span>
+            </div>
+          </article>
+        ))}
+      </div>
+    </>
   );
 }
 
-export default News;
+export default React.memo(Popular, isEqual);
