+++ components/features/detail/content/header-content/index.tsx
