+++ components/features/detail/content/index.tsx
